convert loadgraphval to double

+++ src/common.h
@@ -67,7 +67,7 @@ void scan_loadavg_arg(struct text_object *, const char *);
 void print_loadavg(struct text_object *, char *, int);
 #ifdef X11
 void scan_loadgraph_arg(struct text_object *, const char *);
-uint8_t loadgraphval(struct text_object *);
+double loadgraphval(struct text_object *);
 #endif /* X11 */
 
 uint8_t cpu_percentage(struct text_object *);
